requirements.txt
numpy>=1.26.0
sounddevice>=0.4.6
keyboard>=0.13.5
pyperclip>=1.9.0
huggingface_hub>=0.29.0
transformers>=4.48.0
torch>=2.2.0
faster-whisper>=1.1.1
sherpa-onnx>=1.10.28
python-dotenv>=1.0.1
